✅ run event store in e2e tests
