+++ marketplace-tests/Cargo.toml
@@ -50,6 +50,7 @@ tokio = {version = "1.18.2", features = ["full"]}
 marketplace-core = {path = "../marketplace-core"}
 marketplace-indexer = {path = "../marketplace-indexer"}
 marketplace-infrastructure = {path = "../marketplace-infrastructure"}
+marketplace-event-store = {path = "../marketplace-event-store"}
 
 # Starknet
 starknet = {git = "https://github.com/xJonathanLEI/starknet-rs"}
